ai-chat/backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.0
anthropic==0.49.0
sqlalchemy==2.0.40
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
anyio==4.9.0
